feat: refactor common behaviour of price converters
